+++ hibernate-crud-api/src/main/java/com/thinhle/hibernatecrudapi/rest/StudentRestController.java
@@ -30,33 +30,10 @@ public class StudentRestController {
     @GetMapping("/students/{id}")
     public Student getStudentById(@PathVariable int id) {
         if (id >= students.size() || id < 0) {
-            throw new StudentNotFoundException(String.format("Student %d can't be found", id));
+            throw new StudentNotFoundException(String.format("Student with id %d can't be found", id));
         }
 
         return students.get(id);
-    }
-
-    // Exception Handler
-    @ExceptionHandler
-    public ResponseEntity<StudentErrorResponse> handleException(StudentNotFoundException e) {
-        StudentErrorResponse error = new StudentErrorResponse();
-
-        error.setStatus(HttpStatus.NOT_FOUND.value());
-        error.setMessage(e.getMessage());
-        error.setTimeStamp(System.currentTimeMillis());
-
-        return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<StudentErrorResponse> handleException(Exception e) {
-        StudentErrorResponse error = new StudentErrorResponse();
-
-        error.setStatus(HttpStatus.BAD_REQUEST.value());
-        error.setMessage(e.getMessage());
-        error.setTimeStamp(System.currentTimeMillis());
-
-        return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
     }
 }
